MyApplication/backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
